+++ sklearn_genetic/schedules/base.py
@@ -2,10 +2,10 @@ from abc import ABC
 
 
 class BaseAdapter(ABC):
-    def __init__(self, initial_value, end_value, decay_rate, **kwargs):
+    def __init__(self, initial_value, end_value, adaptive_rate, **kwargs):
         self.initial_value = initial_value
         self.end_value = end_value
-        self.decay_rate = decay_rate
+        self.adaptive_rate = adaptive_rate
         self.current_value = self.initial_value
         self.current_step = 0