+++ sklearn_genetic/schedules/tests/test_schedules.py
@@ -1,17 +1,16 @@
 import pytest
-from ..schedulers import ExponentialAdapter, InverseAdapter
+from ..schedulers import ExponentialAdapter, InverseAdapter, PotentialAdapter
 
 
 @pytest.mark.parametrize(
     "method, params",
     [
-        (
-                ExponentialAdapter,
-                {"initial_value": 0.6, "decay_rate": 0.01, "end_value": 0.4},
-        ),
-        (ExponentialAdapter, {"initial_value": 0.6, "end_value": 0.4, "decay_rate": 0}),
-        (InverseAdapter, {"initial_value": 0.6, "end_value": 0.4, "decay_rate": 0.01}),
-        (InverseAdapter, {"initial_value": 0.6, "end_value": 0.4, "decay_rate": 0}),
+        (ExponentialAdapter,{"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0.01}),
+        (ExponentialAdapter, {"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0}),
+        (InverseAdapter, {"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0.01}),
+        (InverseAdapter, {"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0}),
+        (PotentialAdapter, {"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0.01}),
+        (PotentialAdapter, {"initial_value": 0.6, "end_value": 0.4, "adaptive_rate": 0}),
     ],
 )
 def test_scheduler_decay_limits(method, params):
@@ -26,13 +25,12 @@ def test_scheduler_decay_limits(method, params):
 @pytest.mark.parametrize(
     "method, params",
     [
-        (
-                ExponentialAdapter,
-                {"initial_value": 0.2, "decay_rate": 0.01, "end_value": 0.5},
-        ),
-        (ExponentialAdapter, {"initial_value": 0.2, "end_value": 0.5, "decay_rate": 0}),
-        (InverseAdapter, {"initial_value": 0.2, "end_value": 0.5, "decay_rate": 0.01}),
-        (InverseAdapter, {"initial_value": 0.2, "end_value": 0.5, "decay_rate": 0}),
+        (ExponentialAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0.01}),
+        (ExponentialAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0}),
+        (InverseAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0.01}),
+        (InverseAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0}),
+        (PotentialAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0.01}),
+        (PotentialAdapter, {"initial_value": 0.2, "end_value": 0.5, "adaptive_rate": 0}),
     ],
 )
 def test_scheduler_ascent_limits(method, params):